make subscription domain configurable (#98)
Currently, when a subscription request happens, the subscription server
always uses the primary domain of the `CAPApplication` in the
subscription callback response. The aim of this PR is to make this
configurable by setting the optional annotation
`sme.sap.com/subscription-domain` on the `CAPApplication`. If unset, the
current behaviour will be applied.

Co-authored-by: Christoph Barbian <christoph.barbian@sap.com>
Co-authored-by: Pavan Nayak <pavan.nayak@sap.com>