[FEAT] PDB: workload specific configuration added (#302)
Introduced Pod Disruption Budget (PDB) support for workloads via an
optional `podDisruptionBudget` config in the `DeploymentDetails` struct.
This allows consumers to have better control over workload availability
and deal with disruptions.

Webhook:
The `Selector` field in the PDB spec will be filled automatically and
must not be set by consumers.